fix(widgetbook): split move dialog stories into catalog_part4.dart for repo.part_unit_size (Refs #2751)
PR #2757's earlier slice pushed catalog_part3.dart to 1024 physical lines,
exceeding the 1000-line cap enforced by repo.part_unit_size. Move the
new MoveArmy/MoveFleet helpers, fixtures, and Widgetbook directories
(_moveDialogStoryFrame, _moveArmyStoryFixture, _moveFleetStoryFixture,
moveArmyDialogDirectories, moveFleetDialogDirectories) into a new
catalog_part4.dart and register the part directive in catalog.dart.

The opener button labels are dev-only Widgetbook strings; suppress
repo.app_hardcoded_ui_strings on those two lines via
// ignore: avoid_hardcoded_strings_in_widgets to avoid generating
ARB entries for catalog-only affordances.

Verification:
- dart run tool/check_part_unit_size.dart -> no violations.
- CT_REPO_LINT_INCLUDE_APP=true dart run tool/ct_repo_lint.dart -> all 47 rule(s) passed.
- cd app && flutter analyze lib/widgetbook/ test/move_dialogs_specs_test.dart -> No issues found.
- cd app && flutter test test/move_dialogs_specs_test.dart -> 12/12 pass.
